Add platform developer entry point
This commit adds information for a platform developer on how to obtain
the repo tool, check out the source code with it and build their
first image.

It's not very nice build because I used `include` and this seems to make
it impossible to exchange parts of it in the main index file. But let's
start with something like that first.