Change to "Read the Docs" theme
The theme we use now doesn't have a sidebar with an easy accessible
and expandable tree for the documentation, therefor it is difficult
to read.

The "Read the Docs" theme, which is also used by the Linux Kernel
documentation is much nicer and will make it possible for us to
stop using the includes and start using the index properly.

There is still a lot to do, for example to rewrite the homepage
and to remove all the includes but this step makes it already
much easier to browse and read.