Restructure the index files and add CI articles
After the resturturing work in the blueprint it is now easier to
use it in a project. This patch changes some of the index files
so they would not use include anymore but toctrees instead.

It also adds some articles from the blueprint like those about the
CI and the blueprint itself.