SOTA-System: Document SWUpdate's bootloader requirements
This adds a paragraph to the "Design Decisions" section of the SOTA
System documentation to explain the way SWUpdate interacts with
bootloaders, how it should be used by artifacts and how it can be
extended for new bootloaders. This is both helpful when writing down
new sw-description files for artifacts and when adaptating the SWUpdate
to an unsupported piece of hardware.

Signed-off-by: Florent Revest <revestflo@gmail.com>