Jenkinsfile-deploy: remove building of tags
In some cases tags can start to fail, e.g. if updates are done on
master which breaks legacy. We would need to edit the tags so they
point to the latest on a release branch to fix this, but until we
have a way to do that we will disable building tags.

Signed-off-by: Therese Nordqvist <therese.nordqvist@pelagicore.com>