Renaming Software Factory to Development Handbook
The term Software Factory is from now on referred to Luxoft
Software Factory and PELUX Software Factory is from now on
referred to the PELUX Development Handbook.

The handbook contains blueprints and processes and can be
considered to be part of or to be an instance of the Luxoft
Software Factory.

The landing page has also been restructured.