Add instructions for putting images onto the ARP
Add instructions for putting PELUX images on to the ARP. This includes the
booting from USB, SD-card and EMMC.

The instructions focus on the ARP as an intel target so some information.
Meaning that if PELUX at a later stage decides to support other SMARC
modules, such as ARM based SMARC modules, then additional information
might need to be added.

Signed-off-by: Viktor Sjölind <vsjolind@luxoft.com>