release-process: mention compressed wic images and bmap files
In list of artifacts.

The size of the wic files almost doubled in size when we added software
update and now we also have a /data partition (almost +1GB). An
uncompressed Intel QtAuto wic image is currently 7.83GB while a bz2
compressed one is 1.23GB.

Including .bmap files in the release means it will also be much faster to
flash. We already have instructions in the handbook for how to use them.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>